refactor: avoid circular import
